chore: 개념 학습 페이지 폴더 구조 변경
