+++ app/(afterLogin)/study/layout.tsx
@@ -4,9 +4,9 @@ import { ReactNode } from 'react';
 export default function StudyLayout({ children }: { children: ReactNode }) {
   return (
     <div className='relative mx-auto flex min-h-screen max-w-7xl pt-8'>
-      <div className='sticky h-fit top-1/2 -translate-y-1/2'>
+      <aside className='sticky top-8 h-fit'>
         <ChapterSidebar />
-      </div>
+      </aside>
       <main className='ml-8 flex-1'>{children}</main>
     </div>
   );
